refactor: extract notification routes
